Update README.md
The mistake in the original text is the use of "your favourite" instead of "your favorite". "Favourite" is a British English spelling, while "favorite" is the American English spelling. Additionally, "it" in "you can add it" is ambiguous and could refer to either the missing feature or the experience. To make the sentence clearer, it would be better to specify what "it" refers to, such as "you can add the missing feature and make the experience a beautiful one for all."